requirements.txt
amqp==5.3.1
asgiref==3.8.1
beautifulsoup4==4.14.2
billiard==4.2.2
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
cryptography==46.0.1
dj-database-url==3.0.1
Django==5.2.2
django-allauth==65.11.2
django-ckeditor-5==0.2.18
django-cloudinary-storage==0.3.0
django-debug-toolbar==6.0.0
django-environ==0.12.0
django-js-asset==3.1.2
django-redis==6.0.0
django-unfold==0.67.0
idna==3.10
kombu==5.5.4
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg2==2.9.11
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-magic==0.4.27
redis==7.0.1
requests==2.32.5
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
